Struct tokio_sync::mpsc::block::Block

pub(crate) struct Block<T> {
    start_index: usize,
    next: AtomicPtr<Block<T>>,
    ready_slots: AtomicUsize,
    observed_tail_position: CausalCell<usize>,
    values: Values<T>,
}

A block in a linked list.

Each block in the list can hold up to BLOCK_CAP messages.

Fields
start_index: usize

The start index of this block.

Slots in this block have indices in start_index .. start_index + BLOCK_CAP.

next: AtomicPtr<Block<T>>

The next block in the linked list.

ready_slots: AtomicUsize

Bitfield tracking slots that are ready to have their values consumed.

observed_tail_position: CausalCell<usize>

The observed tail_position value after the block has been passed by block_tail.

values: Values<T>

Array containing values pushed into the block. Values are stored in a continuous array in order to improve cache line behavior when reading. The values must be manually dropped.

Implementations

impl<T> Block<T>

pub(crate) fn new(start_index: usize) -> Block<T>

pub(crate) fn is_at_index(&self, index: usize) -> bool

Returns true if the block matches the given index

pub(crate) fn distance(&self, other_index: usize) -> usize

Returns the number of blocks between self and the block at the specified index.

start_index must represent a block after self.

pub(crate) unsafe fn read(&self, slot_index: usize) -> Option<Read<T>>

Read the value at the given offset.

Returns None if the slot is empty.

Safety

To maintain safety, the caller must ensure:

No concurrent access to the slot.

pub(crate) unsafe fn write(&self, slot_index: usize, value: T)

Write a value to the block at the given offset.

Safety

To maintain safety, the caller must ensure:

The slot is empty.

No concurrent access to the slot.

pub(crate) unsafe fn tx_close(&self)

Signal to the receiver that the sender half of the list is closed.

pub(crate) unsafe fn reclaim(&mut self)

Reset the block to a blank state. This enables reusing blocks in the channel.

Safety

To maintain safety, the caller must ensure:

All slots are empty.

The caller holds a unique pointer to the block.

pub(crate) unsafe fn tx_release(&self, tail_position: usize)

Release the block to the rx half for freeing.

This function is called by the tx half once it can be guaranteed that no more senders will attempt to access the block.

Safety

To maintain safety, the caller must ensure:

The block will no longer be accessed by any sender.

fn set_ready(&self, slot: usize)

Mark a slot as ready

pub(crate) fn is_final(&self) -> bool

Returns true when all slots have their ready bits set.

This indicates that the block is in its final state and will no longer be mutated.

Implementation

The implementation walks each slot checking the ready flag. It might be that it would make more sense to coalesce ready flags as bits in a single atomic cell. However, this could have negative impact on cache behavior as there would be many more mutations to a single slot.

pub(crate) fn observed_tail_position(&self) -> Option<usize>

Returns the observed_tail_position value, if set

pub(crate) fn load_next(&self, ordering: Ordering) -> Option<NonNull<Block<T>>>

Load the next block

pub(crate) unsafe fn try_push(
&self,
block: &mut NonNull<Block<T>>,
ordering: Ordering
) -> Result<(), NonNull<Block<T>>>

Push block as the next block in the link.

Returns Ok if successful, otherwise, a pointer to the next block in the list is returned.

This requires that the next pointer is null.

Ordering

This performs a compare-and-swap on next using AcqRel ordering.

Safety

To maintain safety, the caller must ensure:

block is not freed until it has been removed from the list.

pub(crate) fn grow(&self) -> NonNull<Block<T>>

Grow the Block linked list by allocating and appending a new block.

The next block in the linked list is returned. This may or may not be the one allocated by the function call.

Implementation

It is assumed that self.next is null. A new block is allocated with start_index set to be the next block. A compare-and-swap is performed with AcqRel memory ordering. If the compare-and-swap is successful, the newly allocated block is released to other threads walking the block linked list. If the compare-and-swap fails, the current thread acquires the next block in the linked list, allowing the current thread to access the slots.