Struct tokio_sync::mpsc::Receiver

pub struct Receiver<T> { /* fields omitted */ }

Receive values from the associated Sender.

Instances are created by the channel function.

Implementations

impl<T> Receiver<T>

pub fn close(&mut self)

Closes the receiving half of a channel, without dropping it.

This prevents any further messages from being sent on the channel while still enabling the receiver to drain messages that are buffered.