Module tokio_sync::oneshot

A channel for sending a single message between asynchronous tasks.

Modules

error

Oneshot error types

Structs

Receiver

Receive a value from the associated Sender.

Sender

Sends a value to the associated Receiver.

Functions

channel

Create a new one-shot channel for sending single values across asynchronous tasks.