Module tokio_sync::semaphore

Thread-safe, asynchronous counting semaphore.

A Semaphore instance holds a set of permits. Permits are used to synchronize access to a shared resource.

Before accessing the shared resource, callers acquire a permit from the semaphore. Once the permit is acquired, the caller then enters the critical section. If no permits are available, then acquiring the semaphore returns NotReady. The task is notified once a permit becomes available.

Structs

AcquireError
Error returned by Permit::poll_acquire.

Permit
A semaphore permit

SemState
Semaphore state

Semaphore
Futures-aware semaphore.

TryAcquireError
Error returned by Permit::try_acquire.

WaiterNode
Node used to notify the semaphore waiter when permit is available.

Enums

ErrorKind

NodeState
Waiter node state

PermitState
Permit state

Constants

CLOSED_FLAG

MAX_PERMITS

NUM_FLAG
Flag differentiating between available permits and waiter pointers.

NUM_SHIFT
When representing "numbers", the state has to be shifted this much (to get rid of the flag bit).

Functions

to_try_acquire