Struct tokio_sync::semaphore::Permit

pub struct Permit { /* fields omitted */ }

A semaphore permit

Tracks the lifecycle of a semaphore permit.

An instance of Permit is intended to be used with a single instance of Semaphore. Using a single instance of Permit with multiple semaphore instances will result in unexpected behavior.

Permit does not release the permit back to the semaphore on drop. It is the user’s responsibility to ensure that Permit::release is called before dropping the permit.

Implementations

impl Permit

pub fn new() -> Permit

Create a new Permit.

The permit begins in the “unacquired” state.

Examples

use tokio_sync::semaphore::Permit;

let permit = Permit::new();
assert!(!permit.is_acquired());

pub fn is_acquired(&self) -> bool

Returns true if the permit has been acquired

pub fn poll_acquire(&mut self, semaphore: &Semaphore) -> Poll<(), AcquireError>

Try to acquire the permit. If no permits are available, the current task is notified once a new permit becomes available.

pub fn try_acquire(
&mut self,
semaphore: &Semaphore
) -> Result<(), TryAcquireError>

Try to acquire the permit.

pub fn release(&mut self, semaphore: &Semaphore)

Release a permit back to the semaphore

pub fn forget(&mut self)

Forget the permit without releasing it back to the semaphore.

After calling forget, poll_acquire is able to acquire new permit from the sempahore.

Repeatedly calling forget without associated calls to add_permit will result in the semaphore losing all permits.