Struct tokio_sync::semaphore::SemState

struct SemState(usize);

Semaphore state

The 2 low bits track the modes.

Closed

Full

When not full, the rest of the usize tracks the total number of messages in the channel. When full, the rest of the usize is a pointer to the tail of the “waiting senders” queue.

Tuple Fields
0: usize
Implementations

impl SemState

fn new(permits: usize, stub: &WaiterNode) -> SemState

Returns a new default State value.

fn new_ptr(tail: NonNull<WaiterNode>, closed: bool) -> SemState

Returns a State tracking ptr as the tail of the queue.

fn available_permits(&self) -> usize

Returns the amount of remaining capacity

fn has_available_permits(&self) -> bool

Returns true if the state has permits that can be claimed by a waiter.

fn has_waiter(&self, stub: &WaiterNode) -> bool

fn acquire_permit(&mut self, stub: &WaiterNode) -> bool

Try to acquire a permit

Return

Returns true if the permit was acquired, false otherwise. If false is returned, it can be assumed that State represents the head pointer in the mpsc channel.

fn release_permits(&mut self, permits: usize, stub: &WaiterNode)

Release permits

Returns true if the permits were accepted.

fn is_waiter(&self) -> bool

fn waiter(&self) -> Option<NonNull<WaiterNode>>

Returns the waiter, if one is set.

fn as_ptr(&self) -> *mut WaiterNode

Assumes self represents a pointer

fn set_waiter(&mut self, waiter: NonNull<WaiterNode>)

Set to a pointer to a waiter.

This can only be done from the full state.

fn is_stub(&self, stub: &WaiterNode) -> bool

fn load(cell: &AtomicUsize, ordering: Ordering) -> SemState

Load the state from an AtomicUsize.

fn swap(&self, cell: &AtomicUsize, ordering: Ordering) -> SemState

Swap the values

fn compare_exchange(
&self,
cell: &AtomicUsize,
prev: SemState,
success: Ordering,
failure: Ordering
) -> Result<SemState, SemState>

Compare and exchange the current value into the provided cell

fn fetch_set_closed(cell: &AtomicUsize, ordering: Ordering) -> SemState

fn is_closed(&self) -> bool

fn to_usize(&self) -> usize

Converts the state into a usize representation.