Struct tokio_sync::semaphore::Semaphore

pub struct Semaphore {
    state: AtomicUsize,
    head: CausalCell<NonNull<WaiterNode>>,
    rx_lock: AtomicUsize,
    stub: Box<WaiterNode>,
}

Futures-aware semaphore.

Fields
state: AtomicUsize

Tracks both the waiter queue tail pointer and the number of remaining permits.

head: CausalCell<NonNull<WaiterNode>>

waiter queue head pointer.

rx_lock: AtomicUsize

Coordinates access to the queue head.

stub: Box<WaiterNode>

Stub waiter node used as part of the MPSC channel algorithm.

Implementations

impl Semaphore

pub fn new(permits: usize) -> Semaphore

Creates a new semaphore with the initial number of permits

Panics

Panics if permits is zero.

pub fn available_permits(&self) -> usize

Returns the current number of available permits

fn poll_permit(&self, permit: Option<&mut Permit>) -> Poll<(), AcquireError>

Poll for a permit

pub fn close(&self)

Close the semaphore. This prevents the semaphore from issuing new permits and notifies all pending waiters.

pub fn add_permits(&self, n: usize)

Add n new permits to the semaphore.

fn add_permits_locked(&self, rem: usize, closed: bool)

fn add_permits_locked2(&self, n: usize, closed: bool)

Release a specific amount of permits to the semaphore

This function is called by add_permits after the add lock has been acquired.

fn pop(&self, rem: usize, closed: bool) -> Option<Arc<WaiterNode>>

Pop a waiter

rem represents the remaining number of times the caller will pop. If there are no more waiters to pop, rem is used to set the available permits.

unsafe fn push_stub(&self, closed: bool)

fn stub(&self) -> NonNull<WaiterNode>