Crate tokio_sync

Asynchronous synchronization primitives.

Note: This crate is deprecated in tokio 0.2.x and has been moved into tokio::sync behind the sync feature flag.

This crate provides primitives for synchronizing asynchronous tasks.

Modules

lock
An asynchronous Mutex-like type.

mpsc
A multi-producer, single-consumer queue for sending values across asynchronous tasks.

oneshot
A channel for sending a single message between asynchronous tasks.

semaphore
Thread-safe, asynchronous counting semaphore.

task
Thread-safe task notification primitives.

watch
A single-producer, multi-consumer channel that only retains the last sent value.