Struct tokio_sync::watch::Sender

pub struct Sender<T> { /* fields omitted */ }

Sends values to the associated Receiver.

Instances are created by the channel function.

Methods

impl<T> Sender<T>

pub fn broadcast(&self, value: T) -> Result<(), SendError<T>>

Broadcast a new value via the channel, notifying all receivers.

pub async fn closed<'_>(&'_ mut self)

Completes when all receivers have dropped.

This allows the producer to get notified when interest in the produced values is canceled and immediately stop doing work.